+++ packages/breadboard/src/inspector/types.ts
@@ -4,6 +4,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import {
+  GraphIdentifier,
+  NodeMetadata,
+} from "@google-labs/breadboard-schema/graph.js";
 import { HarnessRunResult, SecretResult } from "../harness/types.js";
 import { GraphLoader } from "../loader/types.js";
 import {
@@ -88,6 +92,11 @@ export type InspectableNode = {
    */
   configuration(): NodeConfiguration;
   /**
+   * Returns metadata for the node.
+   * TODO: Use a friendlier to inspection return type.
+   */
+  metadata(): NodeMetadata;
+  /**
    * Returns the current state of node's ports
    */
   ports(
@@ -114,6 +123,8 @@ export type InspectableEdge = {
    */
   in: string;
 };
+
+export type InspectableSubgraphs = Record<GraphIdentifier, InspectableGraph>;
 
 export type InspectableGraph = {
   /**
@@ -173,6 +184,10 @@ export type InspectableGraph = {
    * output of the `NodeDescriberFunction`.
    */
   describe(): Promise<NodeDescriberResult>;
+  /**
+   * Returns the subgraphs that are embedded in this graph.
+   */
+  graphs(): InspectableSubgraphs;
 };
 
 /**
@@ -644,14 +659,6 @@ export type InspectableRun = {
    * null if not found.
    */
   getEventById(id: EventIdentifier): InspectableRunEvent | null;
-  /**
-   * @deprecated Use `events` instead.
-   */
-  messages: HarnessRunResult[];
-  /**
-   * @deprecated Use `events` instead.
-   */
-  currentNode(position: number): string;
 };
 
 /**