+++ packages/graph-playground/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @google-labs/graph-playground
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies [76da09d]
+- Updated dependencies [938015d]
+- Updated dependencies [182a546]
+  - @google-labs/breadboard@0.15.0
+  - @google-labs/core-kit@0.5.3
+  - @google-labs/graph-integrity@0.0.9
+  - @google-labs/json-kit@0.1.4
+  - @google-labs/node-nursery@0.0.10
+  - @google-labs/palm-kit@0.0.9
+  - @google-labs/pinecone-kit@0.1.7
+  - @google-labs/template-kit@0.2.4
+
 ## 0.0.10
 
 ### Patch Changes